requirements.txt
### REQUIREMENTS FOR MCP23017 AND GPIO MULTIPLEXER ###
RPi.GPIO
smbus2
